requirements.txt
opencv-python
python-vlc
matplotlib
numpy
